Added a dbTEst to make sure it was connected properly, also organized our project a bit
