+++ src/main/java/com/group6/backend/controller/HealthController.java
@@ -1,9 +1,10 @@
-package com.group6.backend;
+package com.group6.backend.controller;
 
 import java.util.Map;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
+
 
 @RestController
 public class HealthController {
