+++ src/main/java/com/group6/backend/model/PartOfSpeech.java
@@ -1,0 +1,16 @@
+package com.group6.backend.model;
+
+// Model defines database tables
+// Part Of Speech file can be used anywhere, good for creating new words using our api
+public enum PartOfSpeech {
+    NOUN,
+    VERB,
+    ADJECTIVE,
+    ADVERB,
+    PRONOUN,
+    PREPOSITION,
+    CONJUNCTION,
+    INTERJECTION,
+    DETERMINER,
+    PHRASE
+}
