+++ src/main/java/com/group6/backend/repo/WordsRepo.java
@@ -1,0 +1,10 @@
+package com.group6.backend.repo;
+
+import org.springframework.stereotype.Repository;
+
+// Repo is basically a database helper, used to query the database
+// Will work on this later
+@Repository
+public interface WordsRepo {
+    
+}
